Fixed links & added link about project templates
Links were broken and didn't have the .md suffix.
Also there was an article about creating simple project templates using add-ins but wasn't yet added to the index file.

Note: the article "Creating a Simple Project template" is not the same as the project templates article. The project templates article only explains the structure, but does not explain how to create a simple project template.