Remove Google Analytics
## Remove Google Analytics

This commit removes the Google Analytics 'functionality' from this page. Following on
from the previous commit, this change also greatly improves the privacy of 
unsuspecting users.

If you must have an analytics system, might I suggest either:
- Plausible: https://plausible.io (offers a 50% discount for FOSS projects)
- Goatcounter: https://goatcounter.com
- Matomo: https://matomo.org
- Fathom: https://usefathom.org

These solutions are all well-tested and open-source. This, of course, means they
can be independently audited for privacy and security. Instead of using a proprietary
nonfree system such as Google Analytics, perhaps we could work together to replace
it with something slightly more ethical?

Furthermore, the majority of privacy features in evergreen browsers *block* Google
Analytics from being run. I can not say the same for solutions such as Plausible,
which suggests they are trusted for their privacy by independent persons.

Thankfully, this change is relatively easy to make, as you have a brilliant
taste in build tools! Jekyll is, thankfully, rather easy to work with.